Factor out AST logic
