+++ src/cli/commands/codegen/route-types.ts
@@ -1,10 +1,8 @@
 import { Option } from "clipanion"
-import { Project, Type, ts, Symbol, TypeFormatFlags } from "ts-morph"
-import path from "node:path"
 import fs from "node:fs/promises"
-import { createRouteMapFromDirectory } from "src/routes/create-route-map-from-directory.js"
 import { BaseCommand } from "src/cli/base-command.js"
 import { ResolvedEdgeSpecConfig } from "src/config/utils.js"
+import { extractRouteSpecsFromAST } from "src/lib/codegen/extract-route-specs-from-ast.js"
 
 export class CodeGenRouteTypes extends BaseCommand {
   static paths = [[`codegen`, `route-types`]]
@@ -15,113 +13,10 @@ export class CodeGenRouteTypes extends BaseCommand {
   })
 
   async run(config: ResolvedEdgeSpecConfig) {
-    const project = new Project({
-      compilerOptions: {
-        declaration: true,
-        noEmit: false,
-        emitDeclarationOnly: true,
-      },
+    const { project, routes, renderType } = await extractRouteSpecsFromAST({
       tsConfigFilePath: config.tsconfigPath,
+      routesDirectory: config.routesDirectory,
     })
-
-    const diagnostics = project.getPreEmitDiagnostics()
-    if (diagnostics.length > 0) {
-      this.context.stderr.write(
-        project.formatDiagnosticsWithColorAndContext(diagnostics)
-      )
-      throw new Error("Type generation failed (existing type errors)")
-    }
-
-    const routeMap = await createRouteMapFromDirectory(config.routesDirectory)
-
-    const nodes = Object.entries(routeMap).map(([route, { relativePath }]) => {
-      const source = project.getSourceFileOrThrow(
-        path.join(config.routesDirectory, relativePath)
-      )
-
-      const defaultExportDeclaration = source
-        .getExportedDeclarations()
-        .get("default")?.[0]
-      if (!defaultExportDeclaration) {
-        console.warn(`No default export found for ${route}`)
-        return
-      }
-
-      const callExpression = defaultExportDeclaration?.getChildrenOfKind(
-        ts.SyntaxKind.CallExpression
-      )[0]
-      if (!callExpression) {
-        return
-      }
-
-      const callSignature = project
-        .getTypeChecker()
-        .getResolvedSignature(callExpression)
-      if (!callSignature) {
-        return
-      }
-
-      const firstParameter = callSignature?.getParameters()?.[0]
-      if (!firstParameter) {
-        return
-      }
-      const parameterType = project
-        .getTypeChecker()
-        .getTypeOfSymbolAtLocation(
-          firstParameter,
-          firstParameter.getValueDeclarationOrThrow()
-        )
-
-      const httpMethodsSymbol = parameterType.getProperty("methods")
-      if (!httpMethodsSymbol) {
-        return
-      }
-
-      const httpMethodLiterals = httpMethodsSymbol
-        .getValueDeclarationOrThrow()
-        .getDescendantsOfKind(ts.SyntaxKind.StringLiteral)
-        .map((d) => d.getLiteralText())
-
-      return {
-        route,
-        httpMethods: httpMethodLiterals,
-        jsonResponseZodOutputType: getZodTypeOfSymbol(
-          project,
-          parameterType.getProperty("jsonResponse")
-        ),
-        jsonBodyZodInputType: getZodTypeOfSymbol(
-          project,
-          parameterType.getProperty("jsonBody")
-        ),
-        commonParamsZodInputType: getZodTypeOfSymbol(
-          project,
-          parameterType.getProperty("commonParams")
-        ),
-        queryParamsZodInputType: getZodTypeOfSymbol(
-          project,
-          parameterType.getProperty("queryParams")
-        ),
-        urlEncodedFormDataZodInputType: getZodTypeOfSymbol(
-          project,
-          parameterType.getProperty("urlEncodedFormData")
-        ),
-      }
-    })
-
-    const filteredNodes = nodes.filter(Boolean) as Exclude<
-      (typeof nodes)[number],
-      undefined
-    >[]
-
-    const renderType = <TType extends ts.Type>(type: Type<TType>) => {
-      return project
-        .getTypeChecker()
-        .compilerObject.typeToString(
-          type.compilerType,
-          undefined,
-          TypeFormatFlags.NoTruncation
-        )
-    }
 
     project.createSourceFile(
       "manifest.ts",
@@ -129,7 +24,7 @@ export class CodeGenRouteTypes extends BaseCommand {
       import {z} from "zod"
 
       export type Routes = {
-${filteredNodes
+${routes
   .map(
     ({
       route,
@@ -181,24 +76,4 @@ ${filteredNodes
       result.getFiles().find((f) => f.filePath.includes("/manifest.d.ts"))!.text
     )
   }
-}
-
-function getZodTypeOfSymbol(project: Project, symbol: Symbol | undefined) {
-  if (!symbol) return undefined
-
-  const outerType = project
-    .getTypeChecker()
-    .getTypeOfSymbolAtLocation(symbol, symbol.getValueDeclarationOrThrow())
-
-  const innerType = outerType.getProperty("_input")
-  if (!innerType) {
-    throw new Error(`${symbol.getName()} must be a zod schema`)
-  }
-
-  return project
-    .getTypeChecker()
-    .getTypeOfSymbolAtLocation(
-      innerType,
-      innerType.getValueDeclarationOrThrow()
-    )
 }
